+++ tools/comfyui/tools/comfyui_img2img.py
@@ -4,12 +4,9 @@ import random
 import uuid
 from copy import deepcopy
 from enum import Enum
-from typing import Any, Generator, Union
+from typing import Any, Generator
 import httpx
 import requests
-import websocket
-from httpx import get, post
-from yarl import URL
 from dify_plugin.entities.tool import (
     ToolInvokeMessage,
     ToolParameter,
@@ -20,7 +17,7 @@ from dify_plugin.errors.tool import ToolProviderCredentialValidationError
 from dify_plugin import Tool
 
 
-from tools.comfyui_client import FileType
+from tools.comfyui_client import ComfyUiClient, FileType
 
 SD_TXT2IMG_OPTIONS = {}
 LORA_NODE = {
@@ -58,6 +55,7 @@ class ComfyuiImg2Img(Tool):
         base_url = self.runtime.credentials.get("base_url", "")
         if not base_url:
             yield self.create_text_message("Please input base_url")
+        self.cli = ComfyUiClient(base_url)
         image_server_url = self.runtime.credentials.get("image_server_url", "")
         if not image_server_url:
             yield self.create_text_message("Please input image_server_url")
@@ -67,7 +65,7 @@ class ComfyuiImg2Img(Tool):
         if not model:
             yield self.create_text_message("Please input model")
             return
-        if model not in self.get_checkpoints():
+        if model not in self.cli.get_checkpoints():
             raise ToolProviderCredentialValidationError(f"model {model} does not exist")
         prompt = tool_parameters.get("prompt", "")
         if not prompt:
@@ -75,7 +73,8 @@ class ComfyuiImg2Img(Tool):
             return
         negative_prompt = tool_parameters.get("negative_prompt", "")
         steps = tool_parameters.get("steps", 20)
-        valid_samplers, valid_schedulers = self.get_sample_methods()
+        valid_samplers = self.cli.get_samplers()
+        valid_schedulers = self.cli.get_schedulers()
         sampler_name = tool_parameters.get("sampler_name", "euler")
         if sampler_name not in valid_samplers:
             raise ToolProviderCredentialValidationError(
@@ -95,12 +94,11 @@ class ComfyuiImg2Img(Tool):
             if image.type != FileType.IMAGE:
                 continue
             blob = httpx.get(image_server_url.rstrip("/") + image.url, timeout=3)
-            files = {
-                "image": (image.filename, blob, image.mime_type),
-                "overwrite": "true",
-            }
-            res = requests.post(str(base_url + "/upload/image"), files=files)
-            image_name = res.json().get("name")
+            image_name = self.cli.post_image(image.filename, blob, image.mime_type)
+            if image_name is None:
+                raise ToolProviderCredentialValidationError(
+                    f"File upload to ComfyUI failed"
+                )
             image_names.append(image_name)
         if len(image_names) == 0:
             yield self.create_text_message("Please input images")
@@ -118,130 +116,6 @@ class ComfyuiImg2Img(Tool):
             scheduler=scheduler,
             cfg=cfg,
         )
-
-    def get_checkpoints(self) -> list[str]:
-        """
-        get checkpoints
-        """
-        try:
-            base_url = self.runtime.credentials.get("base_url", None)
-            if not base_url:
-                return []
-            api_url = str(URL(base_url) / "models" / "checkpoints")
-            response = get(url=api_url, timeout=(2, 10))
-            if response.status_code != 200:
-                return []
-            else:
-                return response.json()
-        except Exception as e:
-            return []
-
-    def get_loras(self) -> list[str]:
-        """
-        get loras
-        """
-        try:
-            base_url = self.runtime.credentials.get("base_url", None)
-            if not base_url:
-                return []
-            api_url = str(URL(base_url) / "models" / "loras")
-            response = get(url=api_url, timeout=(2, 10))
-            if response.status_code != 200:
-                return []
-            else:
-                return response.json()
-        except Exception as e:
-            return []
-
-    def get_sample_methods(self) -> tuple[list[str], list[str]]:
-        """
-        get sample method
-        """
-        try:
-            base_url = self.runtime.credentials.get("base_url", None)
-            if not base_url:
-                return ([], [])
-            api_url = str(URL(base_url) / "object_info" / "KSampler")
-            response = get(url=api_url, timeout=(2, 10))
-            if response.status_code != 200:
-                return ([], [])
-            else:
-                data = response.json()["KSampler"]["input"]["required"]
-                return (data["sampler_name"][0], data["scheduler"][0])
-        except Exception as e:
-            return ([], [])
-
-    def get_history(self, base_url, prompt_id):
-        """
-        get history
-        """
-        url = str(URL(base_url) / "history")
-        respond = get(url, params={"prompt_id": prompt_id}, timeout=(2, 10))
-        return respond.json()
-
-    def download_image(self, base_url, filename, subfolder, folder_type):
-        """
-        download image
-        """
-        url = str(URL(base_url) / "view")
-        response = get(
-            url,
-            params={"filename": filename, "subfolder": subfolder, "type": folder_type},
-            timeout=(2, 10),
-        )
-        return response.content
-
-    def queue_prompt_image(self, base_url, client_id, prompt):
-        """
-        send prompt task and rotate
-        """
-        url = str(URL(base_url) / "prompt")
-        respond = post(
-            url,
-            data=json.dumps({"client_id": client_id, "prompt": prompt}),
-            timeout=(2, 10),
-        )
-        prompt_id = respond.json()["prompt_id"]
-        ws = websocket.WebSocket()
-        if "https" in base_url:
-            ws_url = base_url.replace("https", "ws")
-        else:
-            ws_url = base_url.replace("http", "ws")
-        ws.connect(str(URL(f"{ws_url}") / "ws") + f"?clientId={client_id}", timeout=120)
-        output_images = {}
-        while True:
-            out = ws.recv()
-            if isinstance(out, str):
-                message = json.loads(out)
-                if message["type"] == "executing":
-                    data = message["data"]
-                    if data["node"] is None and data["prompt_id"] == prompt_id:
-                        break
-                elif message["type"] == "status":
-                    data = message["data"]
-                    if data["status"]["exec_info"]["queue_remaining"] == 0 and data.get(
-                        "sid"
-                    ):
-                        break
-            else:
-                continue
-        history = self.get_history(base_url, prompt_id)[prompt_id]
-        for o in history["outputs"]:
-            for node_id in history["outputs"]:
-                node_output = history["outputs"][node_id]
-                if "images" in node_output:
-                    images_output = []
-                    for image in node_output["images"]:
-                        image_data = self.download_image(
-                            base_url,
-                            image["filename"],
-                            image["subfolder"],
-                            image["type"],
-                        )
-                        images_output.append(image_data)
-                    output_images[node_id] = images_output
-        ws.close()
-        return output_images
 
     def img2img(
         self,
@@ -278,7 +152,7 @@ class ComfyuiImg2Img(Tool):
 
         try:
             client_id = str(uuid.uuid4())
-            result = self.queue_prompt_image(base_url, client_id, prompt=draw_options)
+            result = self.cli.queue_prompt_image(client_id, prompt=draw_options)
             image = b""
             for node in result:
                 for img in result[node]:
@@ -309,7 +183,7 @@ class ComfyuiImg2Img(Tool):
         ]
         if self.runtime.credentials:
             try:
-                models = self.get_checkpoints()
+                models = self.cli.get_checkpoints()
                 if len(models) != 0:
                     parameters.append(
                         ToolParameter(
@@ -332,7 +206,7 @@ class ComfyuiImg2Img(Tool):
                             ],
                         )
                     )
-                loras = self.get_loras()
+                loras = self.cli.get_loras()
                 if len(loras) != 0:
                     for n in range(1, 4):
                         parameters.append(
@@ -357,7 +231,8 @@ class ComfyuiImg2Img(Tool):
                                 ],
                             )
                         )
-                (sample_methods, schedulers) = self.get_sample_methods()
+                sample_methods = self.cli.get_samplers()
+                schedulers = self.cli.get_schedulers()
                 if len(sample_methods) != 0:
                     parameters.append(
                         ToolParameter(
